Split up liquidation module tests (#2111)
* Split up liquidation module tests

* Split out getRemainingLiquidatableSizeCapacity
