+++ markets/bfp-market/test/integration/modules/LiquidationModule.liquidatePostion.test.ts
@@ -3,14 +3,12 @@ import assertBn from '@synthetixio/core-utils/utils/assertions/assert-bignumber'
 import assertEvent from '@synthetixio/core-utils/utils/assertions/assert-event';
 import assertRevert from '@synthetixio/core-utils/utils/assertions/assert-revert';
 import Wei, { wei } from '@synthetixio/wei';
-import { BigNumber, Signer, ethers, utils } from 'ethers';
-import { shuffle, times } from 'lodash';
-import forEach from 'mocha-each';
-import { PerpCollateral, bootstrap } from '../../bootstrap';
+import { Signer, ethers, utils } from 'ethers';
+import { shuffle } from 'lodash';
+import { bootstrap } from '../../bootstrap';
 import {
   bn,
   genBootstrap,
-  genNumber,
   genOneOf,
   genOrder,
   genOrderFromSizeDelta,
@@ -21,36 +19,25 @@ import {
 import {
   depositMargin,
   commitAndSettle,
-  commitOrder,
   setMarketConfigurationById,
-  getBlockTimestamp,
   withExplicitEvmMine,
   findEventSafe,
   fastForwardBySec,
   extendContractAbi,
-  ADDRESS0,
   getSusdCollateral,
   sleep,
   setMarketConfiguration,
   setBaseFeePerGas,
-  isSusdCollateral,
 } from '../../helpers';
 import { Market, Trader } from '../../typed';
 import { assertEvents } from '../../assert';
-import {
-  calcFlagReward,
-  calcLiquidationKeeperFee,
-  calcTransactionCostInUsd,
-} from '../../calculations';
-import { PerpMarketConfiguration } from '../../generated/typechain/MarketConfigurationModule';
+import { calcLiquidationKeeperFee, calcTransactionCostInUsd } from '../../calculations';
 
 describe('LiquidationModule', () => {
   const bs = bootstrap(genBootstrap());
   const {
-    pool,
     markets,
     collaterals,
-    collateralsWithoutSusd,
     traders,
     keeper,
     keeper2,
@@ -64,611 +51,6 @@ describe('LiquidationModule', () => {
   beforeEach(restore);
 
   afterEach(async () => await setBaseFeePerGas(1, provider()));
-
-  describe('flagPosition', () => {
-    it('should flag a position with a health factor <= 1', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const orderSide = genSide();
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs)
-      );
-      const order = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 10,
-        desiredSide: orderSide,
-      });
-
-      await commitAndSettle(bs, marketId, trader, order);
-
-      // Price falls/rises between 10% should results in a healthFactor of < 1.
-      //
-      // Whether it goes up or down depends on the side of the order.
-      const newMarketOraclePrice = wei(order.oraclePrice)
-        .mul(orderSide === 1 ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-
-      const { healthFactor } = await BfpMarketProxy.getPositionDigest(trader.accountId, marketId);
-      assertBn.lte(healthFactor, bn(1));
-
-      // set base fee to 0 gwei to avoid rounding errors
-      await setBaseFeePerGas(0, provider());
-      const { flagKeeperReward } = await BfpMarketProxy.getLiquidationFees(
-        trader.accountId,
-        marketId
-      );
-
-      const { receipt } = await withExplicitEvmMine(
-        () => BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-        provider()
-      );
-      const keeperAddress = await keeper().getAddress();
-      await assertEvent(
-        receipt,
-        `PositionFlaggedLiquidation(${trader.accountId}, ${marketId}, "${keeperAddress}", ${flagKeeperReward}, ${newMarketOraclePrice})`,
-        BfpMarketProxy
-      );
-    });
-
-    it('should charge a flag reward that is larger than the notional value', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs, {
-          desiredCollateral: genOneOf(collateralsWithoutSusd()),
-          desiredMarginUsdDepositAmount: 10_000,
-        })
-      );
-      // Create some debt
-      const order1 = await genOrder(bs, market, collateral, collateralDepositAmount);
-
-      await commitAndSettle(bs, marketId, trader, order1);
-
-      // Price falls/rises between 10% should results in a healthFactor of < 1.
-      //
-      // Whether it goes up or down depends on the side of the order.
-      const newMarketOraclePrice = wei(order1.oraclePrice)
-        .mul(order1.sizeDelta.gt(0) ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-      await commitAndSettle(
-        bs,
-        marketId,
-        trader,
-        genOrder(bs, market, collateral, collateralDepositAmount, {
-          desiredSize: order1.sizeDelta.mul(-1),
-        })
-      );
-
-      // Create a small order
-      const order2 = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredSize: wei(100).div(newMarketOraclePrice).toBN(),
-      });
-
-      await commitAndSettle(bs, marketId, trader, order2);
-
-      const { collateralUsd, debtUsd } = await BfpMarketProxy.getAccountDigest(
-        trader.accountId,
-        marketId
-      );
-      // 0 debtUsd and collateral bigger than debtUsd
-      assertBn.gt(debtUsd, 0);
-      assertBn.gt(collateralUsd, debtUsd);
-
-      // Calculate the decrease percentage required to get a collateral value just below the debt
-      const decreasePercentage = wei(collateralUsd)
-        .sub(debtUsd)
-        .div(collateralUsd)
-        .add(0.01)
-        .toNumber();
-      const collateralPrice = await collateral.getPrice();
-
-      // Decrease collateral price and assert that collateral is smaller than debt and that position is liquidatable.
-      await collateral.setPrice(wei(collateralPrice).mul(wei(1).sub(decreasePercentage)).toBN());
-      const { collateralUsd: newCollateralUsd, position } = await BfpMarketProxy.getAccountDigest(
-        trader.accountId,
-        marketId
-      );
-      assertBn.lt(position.healthFactor, 1);
-
-      assertBn.lt(newCollateralUsd, debtUsd);
-      // Make sure collateral is bigger than size so the flag reward is based on collateral
-      assertBn.gt(newCollateralUsd, wei(position.size).mul(newMarketOraclePrice).toBN());
-
-      const baseFeePerGas = await setBaseFeePerGas(0, provider());
-      const { receipt } = await withExplicitEvmMine(
-        () => BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-        provider()
-      );
-      const flagEvent = findEventSafe(receipt, 'PositionFlaggedLiquidation', BfpMarketProxy);
-      const { answer: ethPrice } = await bs.ethOracleNode().agg.latestRoundData();
-
-      const expectedFlagReward = calcFlagReward(
-        ethPrice,
-        baseFeePerGas,
-        wei(order2.sizeDelta.abs()),
-        wei(newMarketOraclePrice),
-        wei(newCollateralUsd),
-        await BfpMarketProxy.getMarketConfiguration(),
-        await BfpMarketProxy.getMarketConfigurationById(marketId)
-      );
-
-      assertBn.equal(flagEvent.args.flagKeeperReward, expectedFlagReward.toBN());
-    });
-
-    it('should charge the larger flagKeeperReward when notional value > collateral value', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs)
-      );
-      const order = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 10,
-      });
-
-      await commitAndSettle(bs, marketId, trader, order);
-
-      // Price falls/rises between 10% should results in a healthFactor of < 1.
-      //
-      // Whether it goes up or down depends on the side of the order.
-      const newMarketOraclePrice = wei(order.oraclePrice)
-        .mul(order.sizeDelta.gt(0) ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-
-      const { healthFactor } = await BfpMarketProxy.getPositionDigest(trader.accountId, marketId);
-      assertBn.lte(healthFactor, bn(1));
-
-      // Grab collateral value before flagging so we can use it to calculate the expected flag reward.
-      const { collateralUsd } = await BfpMarketProxy.getAccountDigest(trader.accountId, marketId);
-
-      const baseFeePerGas = await setBaseFeePerGas(0, provider());
-      const { receipt } = await withExplicitEvmMine(
-        () => BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-        provider()
-      );
-      const flagEvent = findEventSafe(receipt, 'PositionFlaggedLiquidation', BfpMarketProxy);
-      const { answer: ethPrice } = await bs.ethOracleNode().agg.latestRoundData();
-
-      const expectedFlagReward = calcFlagReward(
-        ethPrice,
-        baseFeePerGas,
-        wei(order.sizeDelta.abs()),
-        wei(newMarketOraclePrice),
-        wei(collateralUsd),
-        await BfpMarketProxy.getMarketConfiguration(),
-        await BfpMarketProxy.getMarketConfigurationById(marketId)
-      );
-
-      assertBn.equal(flagEvent.args.flagKeeperReward, expectedFlagReward.toBN());
-    });
-
-    it('should be able to flag when market is in close only', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const orderSide = genSide();
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs)
-      );
-      const order = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 10,
-        desiredSide: orderSide,
-      });
-
-      await commitAndSettle(bs, marketId, trader, order);
-
-      // Price falls/rises between 10% should results in a healthFactor of < 1.
-      //
-      // Whether it goes up or down depends on the side of the order.
-      const newMarketOraclePrice = wei(order.oraclePrice)
-        .mul(orderSide === 1 ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-
-      const { healthFactor } = await BfpMarketProxy.getPositionDigest(trader.accountId, marketId);
-      assertBn.lte(healthFactor, bn(1));
-
-      await setMarketConfigurationById(bs, marketId, { maxMarketSize: 0 });
-      const { maxMarketSize } = await BfpMarketProxy.getMarketConfigurationById(marketId);
-
-      assertBn.equal(maxMarketSize, 0);
-
-      const { receipt } = await withExplicitEvmMine(
-        () => BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-        provider()
-      );
-      await assertEvent(receipt, 'PositionFlaggedLiquidation', BfpMarketProxy);
-    });
-
-    it('should remove any pending orders when present', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const orderSide = genSide();
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs)
-      );
-      const order1 = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 10,
-        desiredSide: orderSide,
-      });
-      await commitAndSettle(bs, marketId, trader, order1);
-
-      // Commit a new order but don't settle.
-      const order2 = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 0.5,
-        desiredSide: orderSide,
-      });
-
-      await commitOrder(bs, marketId, trader, order2);
-
-      const commitmentTime = await getBlockTimestamp(provider());
-
-      // Price moves 10% and results in a healthFactor of < 1.
-      const newMarketOraclePrice = wei(order2.oraclePrice)
-        .mul(orderSide === 1 ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-
-      const { receipt } = await withExplicitEvmMine(
-        () => BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-        provider()
-      );
-
-      // Just assert that flag is triggered, actual values are tested elsewhere
-      await assertEvent(receipt, 'PositionFlaggedLiquidation', BfpMarketProxy);
-      await assertEvent(
-        receipt,
-        `OrderCanceled(${trader.accountId}, ${marketId}, 0, ${commitmentTime})`,
-        BfpMarketProxy
-      );
-    });
-
-    it('should send flagKeeperReward to keeper', async () => {
-      const { BfpMarketProxy, USD } = systems();
-
-      const orderSide = genSide();
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs)
-      );
-      const order1 = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 10,
-        desiredSide: orderSide,
-      });
-      await commitAndSettle(bs, marketId, trader, order1);
-
-      // Commit a new order but don't settle.
-      const order2 = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 0.5,
-        desiredSide: orderSide,
-      });
-      await commitOrder(bs, marketId, trader, order2);
-
-      // Price moves 10% and results in a healthFactor of < 1.
-      const newMarketOraclePrice = wei(order2.oraclePrice)
-        .mul(orderSide === 1 ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-      const flagKeeper = keeper2();
-
-      const { receipt } = await withExplicitEvmMine(
-        () => BfpMarketProxy.connect(flagKeeper).flagPosition(trader.accountId, marketId),
-        provider()
-      );
-      const flagEvent = findEventSafe(receipt, 'PositionFlaggedLiquidation', BfpMarketProxy);
-
-      assertBn.equal(
-        await USD.balanceOf(await flagKeeper.getAddress()),
-        flagEvent.args.flagKeeperReward
-      );
-    });
-
-    it('should reset account debt and update total trader debt', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const { trader, market, marketId, collateral, collateralDepositAmount, collateralPrice } =
-        await depositMargin(
-          bs,
-          genTrader(bs, { desiredCollateral: genOneOf(collateralsWithoutSusd()) }) // use non-sUSD collateral to make sure we accrue some debt (and don't pay it off with sUSD collateral)
-        );
-
-      // Execute two orders to accrue some debt for the trader.
-      const openOrder = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 2,
-      });
-      await commitAndSettle(bs, marketId, trader, openOrder);
-      const newMarketOraclePrice = wei(openOrder.oraclePrice)
-        .mul(openOrder.sizeDelta.gt(0) ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-
-      const closeOrder = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredSize: openOrder.sizeDelta.mul(-1),
-      });
-      const { receipt } = await commitAndSettle(bs, marketId, trader, closeOrder);
-      const closeEvent = findEventSafe(receipt, 'OrderSettled', BfpMarketProxy);
-      const debtInCollateralAmount = wei(closeEvent.args.accountDebt).div(collateralPrice).toBN();
-      // Execute a new order that will create a position that can be flagged for liquidation.
-      const openFlagOrder = await genOrder(
-        bs,
-        market,
-        collateral,
-        collateralDepositAmount.sub(debtInCollateralAmount),
-        {
-          desiredLeverage: 9,
-        }
-      );
-      await commitAndSettle(bs, marketId, trader, openFlagOrder);
-
-      // Price moves 10% and results in a healthFactor of < 1.
-      const newMarketOraclePrice1 = wei(openFlagOrder.oraclePrice)
-        .mul(openFlagOrder.sizeDelta.gt(0) ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice1);
-
-      const { debtUsd: accountDebtBefore } = await BfpMarketProxy.getAccountDigest(
-        trader.accountId,
-        marketId
-      );
-      const { totalTraderDebtUsd: totalTraderDebtUsdBefore } =
-        await BfpMarketProxy.getMarketDigest(marketId);
-
-      assertBn.gt(accountDebtBefore, 0);
-      assertBn.gt(totalTraderDebtUsdBefore, 0);
-
-      await withExplicitEvmMine(
-        () => BfpMarketProxy.connect(keeper2()).flagPosition(trader.accountId, marketId),
-        provider()
-      );
-
-      const { debtUsd: accountDebtAfter } = await BfpMarketProxy.getAccountDigest(
-        trader.accountId,
-        marketId
-      );
-      const { totalTraderDebtUsd: totalTraderDebtUsdAfter } =
-        await BfpMarketProxy.getMarketDigest(marketId);
-
-      assertBn.isZero(accountDebtAfter);
-      assertBn.isZero(totalTraderDebtUsdAfter);
-    });
-
-    it('should remove all margin collateral on flag', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const orderSide = genSide();
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs)
-      );
-      const order1 = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 10,
-        desiredSide: orderSide,
-      });
-      await commitAndSettle(bs, marketId, trader, order1);
-
-      const d1 = await BfpMarketProxy.getMarginDigest(trader.accountId, marketId);
-      assertBn.gt(d1.marginUsd, 0);
-      assertBn.gt(d1.collateralUsd, 0);
-
-      // Commit a new order but don't settle.
-      const order2 = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 0.5,
-        desiredSide: orderSide,
-      });
-      await commitOrder(bs, marketId, trader, order2);
-
-      // Price moves 10% and results in a healthFactor of < 1.
-      const newMarketOraclePrice = wei(order2.oraclePrice)
-        .mul(orderSide === 1 ? 0.9 : 1.1)
-        .toBN();
-      await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-      const flagKeeper = keeper2();
-
-      await BfpMarketProxy.connect(flagKeeper).flagPosition(trader.accountId, marketId);
-
-      // Verify all collateral deposited must be rmrf to market.
-      const d2 = await BfpMarketProxy.getMarginDigest(trader.accountId, marketId);
-      assertBn.isZero(d2.marginUsd);
-      assertBn.isZero(d2.collateralUsd);
-    });
-
-    forEach([
-      ['sUSD', () => getSusdCollateral(collaterals())],
-      ['non-sUSD', () => genOneOf(collateralsWithoutSusd())],
-    ]).it(
-      'should emit all events in correct order (%s)',
-      async (_, getCollateral: () => PerpCollateral) => {
-        const { BfpMarketProxy, Core, SpotMarket } = systems();
-
-        const orderSide = genSide();
-        const { trader, market, marketId, collateralDepositAmount, collateral } =
-          await depositMargin(bs, genTrader(bs, { desiredCollateral: getCollateral() }));
-
-        const order = await genOrder(bs, market, collateral, collateralDepositAmount, {
-          desiredLeverage: 10,
-          desiredSide: orderSide,
-        });
-        await commitAndSettle(bs, marketId, trader, order);
-
-        // Price falls/rises between 10% should results in a healthFactor of < 1.
-        const newMarketOraclePrice = wei(order.oraclePrice)
-          .mul(orderSide === 1 ? 0.9 : 1.1)
-          .toBN();
-        await market.aggregator().mockSetCurrentPrice(newMarketOraclePrice);
-
-        const { healthFactor } = await BfpMarketProxy.getPositionDigest(trader.accountId, marketId);
-        assertBn.lte(healthFactor, bn(1));
-
-        // Set base fee to 0 gwei to avoid rounding errors
-        await setBaseFeePerGas(0, provider());
-
-        const { flagKeeperReward } = await BfpMarketProxy.getLiquidationFees(
-          trader.accountId,
-          marketId
-        );
-        const { receipt } = await withExplicitEvmMine(
-          () => BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-          provider()
-        );
-
-        const blockTime = (await provider().getBlock(receipt.blockNumber)).timestamp;
-        const keeperAddress = await keeper().getAddress();
-        const distributorAddress = collateral.rewardDistributorAddress();
-        const collateralAddress = collateral.synthAddress();
-        const poolCollateralAddress = pool().collateral().address;
-        const poolId = pool().id;
-
-        let expectedEvents: Array<string | RegExp>;
-        if (isSusdCollateral(collateral)) {
-          expectedEvents = [
-            `Transfer("${ADDRESS0}", "${keeperAddress}", ${flagKeeperReward})`,
-            new RegExp(
-              `MarketUsdWithdrawn\\(${marketId}, "${keeperAddress}", ${flagKeeperReward}, "${BfpMarketProxy.address}",`
-            ), // + tail properties omitted.
-            `PositionFlaggedLiquidation(${trader.accountId}, ${marketId}, "${keeperAddress}", ${flagKeeperReward}, ${newMarketOraclePrice})`,
-          ];
-        } else {
-          expectedEvents = [
-            // Withdraw and transfer from Core -> PerpMarket.
-            `Transfer("${Core.address}", "${BfpMarketProxy.address}", ${collateralDepositAmount})`,
-            new RegExp(
-              `MarketCollateralWithdrawn\\(${marketId}, "${collateralAddress}", ${collateralDepositAmount}, "${BfpMarketProxy.address}",`
-            ), // + tail properties omitted.
-            // Transfer from PerpMarket -> RewardDistributor.
-            `Transfer("${BfpMarketProxy.address}", "${distributorAddress}", ${collateralDepositAmount})`,
-            `RewardsDistributed(${poolId}, "${poolCollateralAddress}", "${distributorAddress}", ${collateralDepositAmount}, ${blockTime}, 0)`,
-            // Transfer flag reward from PerpMarket -> Keeper.
-            `Transfer("${ADDRESS0}", "${keeperAddress}", ${flagKeeperReward})`,
-            new RegExp(
-              `MarketUsdWithdrawn\\(${marketId}, "${keeperAddress}", ${flagKeeperReward}, "${BfpMarketProxy.address}",`
-            ), // + tail properties omitted.
-            `PositionFlaggedLiquidation(${trader.accountId}, ${marketId}, "${keeperAddress}", ${flagKeeperReward}, ${newMarketOraclePrice})`,
-          ];
-        }
-
-        // Create a contract that can parse all events emitted.
-        //
-        // This isn't necessary for sUSD collateral but it doesn't affect the correctness of this test.
-        const spotMarketEvents = SpotMarket.interface.format(utils.FormatTypes.full);
-        const contractsWithAllEvents = extendContractAbi(
-          BfpMarketProxy,
-          Core.interface
-            .format(utils.FormatTypes.full)
-            .concat(spotMarketEvents)
-            .concat(['event Transfer(address indexed from, address indexed to, uint256 value)'])
-        );
-        await assertEvents(receipt, expectedEvents, contractsWithAllEvents);
-      }
-    );
-
-    it('should revert when position already flagged', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const orderSide = genSide();
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs)
-      );
-      const order = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 10,
-        desiredSide: orderSide,
-      });
-
-      await commitAndSettle(bs, marketId, trader, order);
-
-      await market.aggregator().mockSetCurrentPrice(
-        wei(order.oraclePrice)
-          .mul(orderSide === 1 ? 0.9 : 1.1)
-          .toBN()
-      );
-
-      // First flag should be successful.
-      await BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId);
-
-      // Second flag should fail because already flagged.
-      await assertRevert(
-        BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-        `PositionFlagged()`,
-        BfpMarketProxy
-      );
-    });
-
-    it('should revert when position health factor > 1', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const orderSide = genSide();
-      const { trader, market, marketId, collateral, collateralDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs)
-      );
-      const order = await genOrder(bs, market, collateral, collateralDepositAmount, {
-        desiredLeverage: 10,
-        desiredSide: orderSide,
-      });
-
-      await commitAndSettle(bs, marketId, trader, order);
-
-      // Position just opened and cannot be liquidated.
-      await assertRevert(
-        BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-        `CannotLiquidatePosition()`,
-        BfpMarketProxy
-      );
-    });
-
-    it('should revert when no open position', async () => {
-      const { BfpMarketProxy } = systems();
-      const { trader, marketId } = await depositMargin(bs, genTrader(bs));
-
-      // Position just opened and cannot be liquidated.
-      await assertRevert(
-        BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, marketId),
-        `PositionNotFound()`,
-        BfpMarketProxy
-      );
-    });
-
-    it('should revert when accountId does not exist', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const { marketId } = await depositMargin(bs, genTrader(bs));
-      const invalidAccountId = 42069;
-
-      await assertRevert(
-        BfpMarketProxy.connect(keeper()).flagPosition(invalidAccountId, marketId),
-        `PositionNotFound()`,
-        BfpMarketProxy
-      );
-    });
-
-    it('should revert when marketId does not exist', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const { trader } = await depositMargin(bs, genTrader(bs));
-      const invalidMarketId = 42069;
-
-      await assertRevert(
-        BfpMarketProxy.connect(keeper()).flagPosition(trader.accountId, invalidMarketId),
-        `MarketNotFound("${invalidMarketId}")`,
-        BfpMarketProxy
-      );
-    });
-
-    describe('distributeRewards', () => {
-      it('should not distribute any rewards when only sUSD collateral');
-
-      it('should only distribute non-sUSD collateral when both available');
-
-      it('should deduct market deposited collateral by distributed amount');
-
-      it('should distribute proportionally across pool collaterals on flag');
-    });
-  });
 
   describe('liquidatePosition', () => {
     const commitAndSettleLiquidatedPosition = async (desiredKeeper: ethers.Signer) => {
@@ -1118,12 +500,12 @@ describe('LiquidationModule', () => {
     it('should update utilization rate after market update during liqudation', async () => {
       const { BfpMarketProxy, Core } = systems();
       const { receipt } = await commitAndSettleLiquidatedPosition(keeper());
-
+      const coreEvents = Core.interface.format(utils.FormatTypes.full) as string[];
       const contractsWithAllEvents = extendContractAbi(
         BfpMarketProxy,
-        Core.interface
-          .format(utils.FormatTypes.full)
-          .concat(['event Transfer(address indexed from, address indexed to, uint256 value)'])
+        coreEvents.concat([
+          'event Transfer(address indexed from, address indexed to, uint256 value)',
+        ])
       );
 
       // Here we check that 'MarketSizeUpdated' is _before_ 'UtilizationRecomputed'.
@@ -1285,85 +667,6 @@ describe('LiquidationModule', () => {
       it('should return the expected liquidationFees on a large open position');
 
       it('should return the expected liquidationFees on a small open position');
-    });
-
-    describe('getRemainingLiquidatableSizeCapacity', () => {
-      const calcMaxLiquidatableCapacity = (
-        makerFee: BigNumber,
-        takerFee: BigNumber,
-        skewScale: BigNumber,
-        liquidationLimitScalar: BigNumber
-      ) => wei(makerFee.add(takerFee)).mul(skewScale).mul(liquidationLimitScalar).toBN();
-
-      describe('maxLiquidatableCapacity', () => {
-        it('should be calculated relative to makerFee/takerFee and skewScale', async () => {
-          const { BfpMarketProxy } = systems();
-          const market = genOneOf(markets());
-          const marketId = market.marketId();
-
-          const liquidationLimitScalar = bn(1);
-          const makerFee = bn(0.0001);
-          const takerFee = bn(0.0001);
-          const skewScale = bn(1_000_000);
-
-          await setMarketConfigurationById(bs, marketId, {
-            liquidationLimitScalar,
-            makerFee,
-            takerFee,
-            skewScale,
-          });
-
-          const { maxLiquidatableCapacity } =
-            await BfpMarketProxy.getRemainingLiquidatableSizeCapacity(marketId);
-          const expectedMaxLiquidatableCapacity = calcMaxLiquidatableCapacity(
-            makerFee,
-            takerFee,
-            skewScale,
-            liquidationLimitScalar
-          );
-          assertBn.equal(maxLiquidatableCapacity, expectedMaxLiquidatableCapacity);
-        });
-
-        /* Randomly test 10 scalars between 0 and 1 (inclusive, always include boundaries). */
-        forEach([0, ...times(8).map(() => genNumber(0.1, 0.9)), 1]).it(
-          `should scale with liquidationLimitScalar of '%0.5f'`,
-          async (scalar: number) => {
-            const { BfpMarketProxy } = systems();
-            const market = genOneOf(markets());
-            const marketId = market.marketId();
-
-            const makerFee = bn(genNumber(0.0001, 0.0005));
-            const takerFee = bn(genNumber(0.0006, 0.001));
-            const skewScale = bn(1_000_000);
-
-            const liquidationLimitScalar = bn(scalar);
-            await setMarketConfigurationById(bs, marketId, {
-              liquidationLimitScalar,
-              makerFee,
-              takerFee,
-              skewScale,
-            });
-
-            const { maxLiquidatableCapacity } =
-              await BfpMarketProxy.getRemainingLiquidatableSizeCapacity(marketId);
-            const expectedMaxLiquidatableCapacity = calcMaxLiquidatableCapacity(
-              makerFee,
-              takerFee,
-              skewScale,
-              liquidationLimitScalar
-            );
-            assertBn.equal(maxLiquidatableCapacity, expectedMaxLiquidatableCapacity);
-          }
-        );
-      });
-
-      describe('remainingCapacity', () => {
-        it('should be updated to reflect a recent liquidation');
-
-        it(
-          'should be zero when cap has been met and an endorsed liquidator pushes past cap for window'
-        );
-      });
     });
 
     describe('{partialLiquidation,liquidationCapacity,liqKeeperFee}', () => {
@@ -2469,352 +1772,6 @@ describe('LiquidationModule', () => {
       it('should revert when liq cap has been met and not endorsed (concrete)');
 
       it('should revert when pd is below maxPd but liquidation happens in the same block');
-    });
-  });
-
-  describe('getLiquidationMarginUsd', () => {
-    const calcImrAndMmr = (size: Wei, marketConfig: PerpMarketConfiguration.DataStructOutput) => {
-      const imr = Wei.min(
-        size
-          .abs()
-          .div(marketConfig.skewScale)
-          .mul(marketConfig.incrementalMarginScalar)
-          .add(marketConfig.minMarginRatio),
-        wei(marketConfig.maxInitialMarginRatio)
-      );
-      const mmr = imr.mul(marketConfig.maintenanceMarginScalar);
-      return { imr, mmr };
-    };
-
-    it('should revert when invalid marketId', async () => {
-      const { BfpMarketProxy } = systems();
-      const invalidMarketId = 42069;
-
-      await assertRevert(
-        BfpMarketProxy.getLiquidationMarginUsd(0, invalidMarketId, 0),
-        `MarketNotFound("${invalidMarketId}")`,
-        BfpMarketProxy
-      );
-    });
-
-    it('should revert when invalid accountId', async () => {
-      const { BfpMarketProxy } = systems();
-      const { marketId } = await depositMargin(bs, genTrader(bs));
-      const invalidAccountId = 42069;
-      await assertRevert(
-        BfpMarketProxy.getLiquidationMarginUsd(invalidAccountId, marketId, 0),
-        `AccountNotFound("${invalidAccountId}")`,
-        BfpMarketProxy
-      );
-    });
-
-    it('should return 0 when no position found', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const { trader, marketId } = await depositMargin(bs, genTrader(bs));
-      const { mm, im } = await BfpMarketProxy.getLiquidationMarginUsd(
-        trader.accountId,
-        marketId,
-        0
-      );
-
-      assertBn.isZero(mm);
-      assertBn.isZero(im);
-    });
-
-    it('should calculate IM and MM for existing position', async () => {
-      const { BfpMarketProxy } = systems();
-      const {
-        market,
-        collateral,
-        collateralDepositAmount,
-        marketId,
-        trader,
-        marginUsdDepositAmount,
-      } = await depositMargin(bs, genTrader(bs));
-      const order = await genOrder(bs, market, collateral, collateralDepositAmount);
-      await commitAndSettle(bs, marketId, trader, order);
-
-      const baseFeePerGas = await setBaseFeePerGas(0, provider());
-
-      const { mm, im } = await BfpMarketProxy.getLiquidationMarginUsd(
-        trader.accountId,
-        marketId,
-        0
-      );
-
-      // Data for calcs.
-      const { maxLiquidatableCapacity } =
-        await BfpMarketProxy.getRemainingLiquidatableSizeCapacity(marketId);
-      const { answer: ethPrice } = await bs.ethOracleNode().agg.latestRoundData();
-      const absSize = wei(order.sizeDelta).abs();
-      const notional = absSize.mul(order.oraclePrice);
-      const globalConfig = await BfpMarketProxy.getMarketConfiguration();
-      const marketConfig = await BfpMarketProxy.getMarketConfigurationById(marketId);
-
-      const flagReward = calcFlagReward(
-        ethPrice,
-        baseFeePerGas,
-        absSize,
-        wei(order.oraclePrice),
-        wei(marginUsdDepositAmount),
-        globalConfig,
-        marketConfig
-      );
-      const liqReward = calcLiquidationKeeperFee(
-        ethPrice,
-        baseFeePerGas,
-        absSize,
-        wei(maxLiquidatableCapacity),
-        globalConfig
-      );
-      const { imr, mmr } = calcImrAndMmr(absSize, marketConfig);
-
-      const expectedIm = notional
-        .mul(imr)
-        .add(marketConfig.minMarginUsd)
-        .add(liqReward)
-        .add(flagReward);
-      const expectedMm = notional
-        .mul(mmr)
-        .add(marketConfig.minMarginUsd)
-        .add(liqReward)
-        .add(flagReward);
-
-      assertBn.near(im, expectedIm.toBN(), bn(0.000001));
-      assertBn.near(mm, expectedMm.toBN(), bn(0.000001));
-    });
-
-    it('should calculate IM and MM for new position', async () => {
-      const { BfpMarketProxy } = systems();
-      const { market, collateralDepositAmount, marketId, trader, marginUsdDepositAmount } =
-        await depositMargin(bs, genTrader(bs));
-
-      const { answer: marketPrice } = await market.aggregator().latestRoundData();
-      const desiredLeverage = genNumber(-5, 5);
-      const desiredSize = wei(collateralDepositAmount).div(marketPrice).mul(desiredLeverage).toBN();
-      const baseFeePerGas = await setBaseFeePerGas(0, provider());
-      const { mm, im } = await BfpMarketProxy.getLiquidationMarginUsd(
-        trader.accountId,
-        marketId,
-        desiredSize
-      );
-
-      // Data for calcs.
-      const { maxLiquidatableCapacity } =
-        await BfpMarketProxy.getRemainingLiquidatableSizeCapacity(marketId);
-      const { answer: ethPrice } = await bs.ethOracleNode().agg.latestRoundData();
-      const absSize = wei(desiredSize).abs();
-      const notional = absSize.mul(marketPrice);
-      const globalConfig = await BfpMarketProxy.getMarketConfiguration();
-      const marketConfig = await BfpMarketProxy.getMarketConfigurationById(marketId);
-
-      const flagReward = calcFlagReward(
-        ethPrice,
-        baseFeePerGas,
-        absSize,
-        wei(marketPrice),
-        wei(marginUsdDepositAmount),
-        globalConfig,
-        marketConfig
-      );
-      const liqReward = calcLiquidationKeeperFee(
-        ethPrice,
-        baseFeePerGas,
-        absSize,
-        wei(maxLiquidatableCapacity),
-        globalConfig
-      );
-      const { imr, mmr } = calcImrAndMmr(absSize, marketConfig);
-
-      const expectedIm = notional
-        .mul(imr)
-        .add(marketConfig.minMarginUsd)
-        .add(liqReward)
-        .add(flagReward);
-      const expectedMm = notional
-        .mul(mmr)
-        .add(marketConfig.minMarginUsd)
-        .add(liqReward)
-        .add(flagReward);
-
-      assertBn.near(im, expectedIm.toBN(), bn(0.000001));
-      assertBn.near(mm, expectedMm.toBN(), bn(0.000001));
-    });
-
-    it('should calculate IM and MM for modifying a position', async () => {
-      const { BfpMarketProxy } = systems();
-      const {
-        market,
-        collateral,
-        collateralDepositAmount,
-        marketId,
-        trader,
-        marginUsdDepositAmount,
-      } = await depositMargin(bs, genTrader(bs));
-      const order = await genOrder(bs, market, collateral, collateralDepositAmount);
-      await commitAndSettle(bs, marketId, trader, order);
-
-      const desiredSizeDelta = bn(genNumber(-1, 1));
-      const baseFeePerGas = await setBaseFeePerGas(0, provider());
-      const { mm, im } = await BfpMarketProxy.getLiquidationMarginUsd(
-        trader.accountId,
-        marketId,
-        desiredSizeDelta
-      );
-
-      // Data for calcs
-      const { maxLiquidatableCapacity } =
-        await BfpMarketProxy.getRemainingLiquidatableSizeCapacity(marketId);
-      const { answer: ethPrice } = await bs.ethOracleNode().agg.latestRoundData();
-      const { answer: marketPrice } = await market.aggregator().latestRoundData();
-
-      const size = order.sizeDelta.add(desiredSizeDelta);
-      const absSize = wei(size).abs();
-      const notional = absSize.mul(marketPrice);
-      const globalConfig = await BfpMarketProxy.getMarketConfiguration();
-      const marketConfig = await BfpMarketProxy.getMarketConfigurationById(marketId);
-
-      const flagReward = calcFlagReward(
-        ethPrice,
-        baseFeePerGas,
-        absSize,
-        wei(marketPrice),
-        wei(marginUsdDepositAmount),
-        globalConfig,
-        marketConfig
-      );
-      const liqReward = calcLiquidationKeeperFee(
-        ethPrice,
-        baseFeePerGas,
-        absSize,
-        wei(maxLiquidatableCapacity),
-        globalConfig
-      );
-      const { imr, mmr } = calcImrAndMmr(absSize, marketConfig);
-
-      const expectedIm = notional
-        .mul(imr)
-        .add(marketConfig.minMarginUsd)
-        .add(liqReward)
-        .add(flagReward);
-      const expectedMm = notional
-        .mul(mmr)
-        .add(marketConfig.minMarginUsd)
-        .add(liqReward)
-        .add(flagReward);
-
-      assertBn.near(im, expectedIm.toBN(), bn(0.000001));
-      assertBn.near(mm, expectedMm.toBN(), bn(0.000001));
-    });
-
-    it('should cap IMR (and hence IM) by the maxInitialMarginRatio (concrete)', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const collateral = genOneOf(collateralsWithoutSusd());
-      await collateral.setPrice(bn(50_000));
-
-      const { market, marketId, trader, marginUsdDepositAmount } = await depositMargin(
-        bs,
-        genTrader(bs, {
-          desiredMarginUsdDepositAmount: 3_000_000_000, // 3b position
-          desiredCollateral: collateral,
-        })
-      );
-
-      const { answer: marketPrice } = await market.aggregator().latestRoundData();
-
-      const desiredSizeDelta = wei(marginUsdDepositAmount).div(wei(marketPrice)).neg().toBN(); // 1x short
-
-      const baseFeePerGas = await setBaseFeePerGas(0, provider());
-
-      const { im } = await BfpMarketProxy.getLiquidationMarginUsd(
-        trader.accountId,
-        marketId,
-        desiredSizeDelta
-      );
-
-      // Data for calcs
-      const marketConfig = await BfpMarketProxy.getMarketConfigurationById(marketId);
-      const { answer: ethPrice } = await bs.ethOracleNode().agg.latestRoundData();
-      const globalConfig = await BfpMarketProxy.getMarketConfiguration();
-      const { maxLiquidatableCapacity } =
-        await BfpMarketProxy.getRemainingLiquidatableSizeCapacity(marketId);
-
-      const size = wei(desiredSizeDelta).abs();
-      const notional = size.mul(marketPrice);
-
-      const flagReward = calcFlagReward(
-        ethPrice,
-        baseFeePerGas,
-        size,
-        wei(marketPrice),
-        wei(marginUsdDepositAmount),
-        globalConfig,
-        marketConfig
-      );
-      const liqReward = calcLiquidationKeeperFee(
-        ethPrice,
-        baseFeePerGas,
-        size,
-        wei(maxLiquidatableCapacity),
-        globalConfig
-      );
-      // Expect the IMR to be at maxInitialMarginRatio cap.
-      const imr = wei(marketConfig.maxInitialMarginRatio);
-      const expectedIm = notional
-        .mul(imr)
-        .add(wei(marketConfig.minMarginUsd))
-        .add(liqReward)
-        .add(flagReward);
-      console.log({
-        flagReward: flagReward.toNumber(),
-        liqReward: liqReward.toNumber(),
-        im: wei(im).toNumber(),
-        expectedIm: expectedIm.toNumber(),
-        diff: wei(im).sub(expectedIm).toNumber(),
-      });
-      assertBn.equal(im, expectedIm.toBN());
-    });
-
-    it('should allow a position to always 1x even when extremely large', async () => {
-      const { BfpMarketProxy } = systems();
-
-      const market = genOneOf(markets());
-      const marketId = market.marketId();
-      const { answer: marketPrice } = await market.aggregator().latestRoundData();
-
-      const trader = genOneOf(traders());
-
-      // 1M per token to avoid exceeding maxAllowable.
-      const collateral = genOneOf(collateralsWithoutSusd());
-      await collateral.setPrice(bn(1_000_000));
-
-      let accumulatedDepositUsd = wei(0);
-
-      for (let i = 0; i < 10; i++) {
-        const desiredMarginUsdDepositAmount = genNumber(420_000_000, 690_000_000);
-        const { marginUsdDepositAmount } = await depositMargin(
-          bs,
-          genTrader(bs, {
-            desiredMarginUsdDepositAmount,
-            desiredCollateral: collateral,
-            desiredMarket: market,
-            desiredTrader: trader,
-          })
-        );
-        accumulatedDepositUsd = accumulatedDepositUsd.add(marginUsdDepositAmount);
-
-        const desiredSizeDelta = wei(accumulatedDepositUsd).div(wei(marketPrice)).neg().toBN(); // 1x short
-        const { im } = await BfpMarketProxy.getLiquidationMarginUsd(
-          trader.accountId,
-          marketId,
-          desiredSizeDelta
-        );
-
-        // Expect the IM to always be below the total deposited margin.
-        assertBn.lt(im, accumulatedDepositUsd.toBN());
-      }
     });
   });
 });
